+++ spec/spec_helper.rb
@@ -6,7 +6,7 @@ SimpleCov.start do
   add_filter "/rspec/"
 end
 
-require "spicery/spec_helper"
+require "conjunction/spec_helper"
 require "shoulda-matchers"
 
 require "active_record"