+++ lib/flow/operation/accessors.rb
@@ -1,14 +1,16 @@
 # frozen_string_literal: true
 
+require "set"
+
 module Flow
   module Operation
     module Accessors
       extend ActiveSupport::Concern
 
       included do
-        class_attribute :_state_readers, instance_writer: false, default: []
-        class_attribute :_state_writers, instance_writer: false, default: []
-        class_attribute :_state_accessors, instance_writer: false, default: []
+        class_attribute :_state_readers, instance_writer: false, default: Set.new
+        class_attribute :_state_writers, instance_writer: false, default: Set.new
+        class_attribute :_state_accessors, instance_writer: false, default: Set.new
       end
 
       module ClassMethods
@@ -34,23 +36,13 @@ module Flow
         private
 
         def _add_state_reader_tracker(name)
-          return false if _state_readers.include?(name)
-
-          _add_state_accessor_tracker(name) if _state_writers.include?(name)
+          _state_accessors << name if _state_writers.include?(name)
           _state_readers << name
         end
 
         def _add_state_writer_tracker(name)
-          return false if _state_writers.include?(name)
-
-          _add_state_accessor_tracker(name) if _state_readers.include?(name)
+          _state_accessors << name if _state_readers.include?(name)
           _state_writers << name
-        end
-
-        def _add_state_accessor_tracker(name)
-          return if _state_accessors.include?(name)
-
-          _state_accessors << name
         end
 
         def inherited(base)