+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming
 
+## 0.11.0
+
+*Release Date*: 2021/8/31
+
 - Add `prefix` option to state accessor methods ([#203](https://github.com/Freshly/flow/pull/203), [#206](https://github.com/Freshly/flow/pull/206))
 - Use Set instead of Array for accessor trackers ([#205](https://github.com/Freshly/flow/pull/205))
 - Allow multiple state accessor definitions at once ([#207](https://github.com/Freshly/flow/pull/207))